+++ src/app/(basic)/articles/_feature/article/ArticleCard.tsx
@@ -1,5 +1,5 @@
 import { ExternalImage } from '@/src/component/Image/ExternalImage'
-import { Article } from '@/src/app/(basic)/articles/api/route'
+import { Article } from '@/src/app/(basic)/articles/api/constant'
 
 type Props = {
   article: Pick<Article, 'title' | 'category' | 'publishedAt' | 'url' | 'image'>
